+++ pkgs/native_assets_builder/test/build_runner/build_runner_dry_run_test.dart
@@ -5,7 +5,7 @@
 import 'dart:io';
 
 import 'package:file_testing/file_testing.dart';
-import 'package:native_assets_cli/native_assets_cli.dart';
+import 'package:native_assets_cli/native_assets_cli_internal.dart';
 import 'package:test/test.dart';
 
 import '../helpers.dart';