+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 - Added [example/use_dart_api/](example/use_dart_api/) detailing how to use
   `dart_api_dl.h` from the Dart SDK in native code.
-- **Breaking change** Moved code not used in `build.dart` to
-  `package:native_assets_builder`.
+- **Breaking change** Removed all code not used in `build.dart` scripts out of
+  the public API.
+
 
 
 ## 0.3.2